requirements.txt
accelerate
bitsandbytes
diffusers==0.31.0
transformers>=4.45.2
huggingface_hub
hf_transfer>=0.1.8
peft>=0.12.0
decord>=0.6.0
wandb
pandas
torchao>=0.5.0
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
deepspeed
opencv-python
imageio
easydict
xfuser==0.4.2
ray==2.44.0
logfire==3.11.0
pyglet==2.1.3
wmk==0.38